refactor: Spacing 제거
