--- src/components/common/Spacing/Spacing.tsx
@@ -1,7 +1,0 @@
-interface Props {
-  height: number
-}
-
-export default function Spacing({ height }: Props) {
-  return <div style={{ height }} />
-}
