--- src/components/common/Spacing/index.ts
@@ -1,3 +1,0 @@
-import Spacing from './Spacing'
-
-export default Spacing
