+++ src/components/memory/Post/Post.tsx
@@ -1,8 +1,8 @@
 import type { MouseEvent } from 'react'
 
 import Border from '@/components/common/Border/Border'
-import Spacing from '@/components/common/Spacing'
 import { emojies, voices } from '@/mocks/data/reaction'
+import { flexGap } from '@/styles/common'
 import type { IMember } from '@/types'
 import EmojiList from '../../reaction/EmojiList'
 import VoiceList from '../../reaction/VoiceList'
@@ -23,14 +23,12 @@ export default function Post(props: Props) {
   }
   return (
     <>
-      <Border />
-      <Spacing height={10} />
-      <PostContent {...props} />
-      <Spacing height={10} />
-      <EmojiList reactions={emojies} onClick={handleReactionClick} />
-      <Spacing height={10} />
-      <VoiceList reactions={voices} onClick={handleReactionClick} />
-      <Spacing height={20} />
+      <div css={[flexGap(10), { marginBottom: '20px' }]}>
+        <Border />
+        <PostContent {...props} />
+        <EmojiList reactions={emojies} onClick={handleReactionClick} />
+        <VoiceList reactions={voices} onClick={handleReactionClick} />
+      </div>
     </>
   )
 }
