+++ src/pages/SignupPage.tsx
@@ -4,10 +4,10 @@ import { useNavigate } from 'react-router-dom'
 import BottomButton from '@/components/common/BottomButton'
 import DateInputField from '@/components/common/DateInputField'
 import InputField from '@/components/common/InputField'
-import Spacing from '@/components/common/Spacing'
 import { MEMBERS } from '@/consts/SSAFY_12_MEMBERS'
 import { ROUTES } from '@/routes/ROUTES'
 import { useUserStore } from '@/store/userStore'
+import { flexGap } from '@/styles/common'
 import { signupTitleStyle } from '@/styles/signupTitle'
 import { dateToId } from '@/utils/date'
 
@@ -29,33 +29,29 @@ export default function SignupPage() {
   }, [name, password, birthDate])
 
   return (
-    <>
+    <div css={flexGap(8)}>
       <div css={signupTitleStyle}>
         <h1>당신은 누구인가요?</h1>
         <p>지금은 싸피 12기 12반만 가입할 수 있어요.</p>
       </div>
-      <Spacing height={8} />
       <InputField
         label='이름'
         onChange={(v: string) => {
           setName(v)
         }}
       />
-      <Spacing height={8} />
       <DateInputField
         label='생년월일'
         onChange={date => {
           setBirthDate(date)
         }}
       />
-      <Spacing height={8} />
       <InputField
         label='암호 - 우리반 반장님이 뱉은 주스는?'
         onChange={(v: string) => {
           setPassword(v)
         }}
       />
-      <Spacing height={8} />
       <BottomButton
         label='입장하기'
         type={isValid ? 'primary' : 'disabled'}
@@ -64,6 +60,6 @@ export default function SignupPage() {
           navigate(ROUTES.MEMORY_LIST)
         }}
       />
-    </>
+    </div>
   )
 }
